Move all analysis csv reading routines and forms to their own module (keep admin.py cleaner)
